Version 0.9.16
Compatibility with the T-vK ESP32-BLE-Keyboard because of recent changes in the library means that it is no longer needed to use my fork of the library. This also enables usage with NimBLE if people want to. Use the USE_NIMBLE definition and make sure to have the latest Arduino NimBLE installed.